Add MessageHandler interface
The MessageHandler interface combines p2p message validation and handling into
one interface. Most of the handlers already did rely on the validation.

Also we stop using generics for this part of the code. I did try to use a generic
interface definition, but ran into problems with go's type inference.

Anyway, it's probably better to operate on a unified type like p2pmsg.Message
instead of handcuffing ourselves with static types (e.g. look at
dummyMessageHandler in host.go and compare it to what we previously had).
