+++ rolling-shutter/p2pnode/host.go
@@ -10,11 +10,28 @@ import (
 	"github.com/shutter-network/rolling-shutter/rolling-shutter/p2pmsg"
 )
 
-type P2PNode struct {
-	p2p *p2p.P2PHandler
+// dummyMessageHandler validates all p2p messages and emits a log message for each p2p message.
+type dummyMessageHandler struct{}
+
+func (dummyMessageHandler) ValidateMessage(_ context.Context, _ p2pmsg.Message) (bool, error) {
+	return true, nil
 }
 
-func New(config Config) *P2PNode {
+func (dummyMessageHandler) HandleMessage(_ context.Context, msg p2pmsg.Message) ([]p2pmsg.Message, error) {
+	log.Info().Str("message", msg.String()).Msg("received message")
+	return nil, nil
+}
+
+func (dummyMessageHandler) MessagePrototypes() []p2pmsg.Message {
+	return []p2pmsg.Message{
+		&p2pmsg.DecryptionKeyShare{},
+		&p2pmsg.DecryptionKey{},
+		&p2pmsg.DecryptionTrigger{},
+		&p2pmsg.EonPublicKey{},
+	}
+}
+
+func New(config Config) service.Service {
 	p2pHandler := p2p.New(
 		p2p.Config{
 			ListenAddrs:     config.ListenAddresses,
@@ -22,63 +39,8 @@ func New(config Config) *P2PNode {
 			PrivKey:         config.PrivateKey,
 			Environment:     config.Environment,
 			IsBootstrapNode: true,
-		})
-	return &P2PNode{
-		p2p: p2pHandler,
-	}
-}
-
-func (p *P2PNode) validateDecryptionTrigger(_ context.Context, _ *p2pmsg.DecryptionTrigger) (bool, error) {
-	return true, nil
-}
-
-func (p *P2PNode) handleDecryptionTrigger(_ context.Context, msg *p2pmsg.DecryptionTrigger) ([]p2pmsg.Message, error) {
-	msgs := []p2pmsg.Message{}
-	log.Info().Str("message", msg.String()).Msg("received message")
-	return msgs, nil
-}
-
-func (p *P2PNode) validateDecryptionKeyShare(_ context.Context, _ *p2pmsg.DecryptionKeyShare) (bool, error) {
-	return true, nil
-}
-
-func (p *P2PNode) handleDecryptionKeyShare(_ context.Context, msg *p2pmsg.DecryptionKeyShare) ([]p2pmsg.Message, error) {
-	msgs := []p2pmsg.Message{}
-	log.Info().Str("message", msg.String()).Msg("received message")
-	return msgs, nil
-}
-
-func (p *P2PNode) validateEonPublicKey(_ context.Context, _ *p2pmsg.EonPublicKey) (bool, error) {
-	return true, nil
-}
-
-func (p *P2PNode) handleEonPublicKey(_ context.Context, msg *p2pmsg.EonPublicKey) ([]p2pmsg.Message, error) {
-	msgs := []p2pmsg.Message{}
-	log.Info().Str("message", msg.String()).Msg("received message")
-	return msgs, nil
-}
-
-func (p *P2PNode) validateDecryptionKey(_ context.Context, _ *p2pmsg.DecryptionKey) (bool, error) {
-	return true, nil
-}
-
-func (p *P2PNode) handleDecryptionKey(_ context.Context, msg *p2pmsg.DecryptionKey) ([]p2pmsg.Message, error) {
-	msgs := []p2pmsg.Message{}
-	log.Info().Str("message", msg.String()).Msg("received message")
-	return msgs, nil
-}
-
-func (p *P2PNode) Start(ctx context.Context, runner service.Runner) error {
-	p2p.AddValidator(p.p2p, p.validateDecryptionKey)
-	p2p.AddHandlerFunc(p.p2p, p.handleDecryptionKey)
-
-	p2p.AddValidator(p.p2p, p.validateDecryptionTrigger)
-	p2p.AddHandlerFunc(p.p2p, p.handleDecryptionTrigger)
-
-	p2p.AddValidator(p.p2p, p.validateDecryptionKeyShare)
-	p2p.AddHandlerFunc(p.p2p, p.handleDecryptionKeyShare)
-
-	p2p.AddValidator(p.p2p, p.validateEonPublicKey)
-	p2p.AddHandlerFunc(p.p2p, p.handleEonPublicKey)
-	return p.p2p.Start(ctx, runner)
+		},
+	)
+	p2pHandler.AddMessageHandler(dummyMessageHandler{})
+	return p2pHandler
 }
